Add automated creation of API docs using Sphinx.
The API documentation is automatically created using Sphinx with
the ReadTheDocs theme. Call structures are automatically obtained
from the pyHiChi.so binary and the information is compiled into a
doctree that is turned into html. Build the docs by running the
build_docs.sh script, while standing in the docs/ folder. The docs
can be read by opening docs/build/html/index.html in your browser.

In order to improve the API docs, we should populated all docstrings.
This is done in the pyHiChi.cpp file where all the bindings are defined.

This should later be changed, as the build scripts for the docs
technically doesn't belong in the repo. Should strive to host the API docs
on readthedocs.io instead.

- Added Sphinx (sphinx) and the ReadTheDocs theme (sphinx_rtd_theme)
  to requirements.txt.
- Initialized the docs by running <sphinx-quickstart> and making the
  appropriate changes in docs/source/index.rst and docs/source/conf.py
- Created a script for building the docs: docs/build_docs.sh.